pal Cesar

+++ Skip_Android/app/src/main/java/com/simplife/skip/models/Parada.kt
@@ -4,8 +4,8 @@ import java.io.Serializable
 
 class Parada (
     var ubicacion: String,
-    var latitud: Float,
-    var longitud: Float
+    var latitud: Double,
+    var longitud: Double
 
 ) :Serializable {
     override fun toString(): String {
