+++ Skip_Backend/src/main/java/com/simplife/skip/repositories/ParadaRepository.java
@@ -7,5 +7,5 @@ import org.springframework.data.jpa.repository.Query;
 public interface ParadaRepository extends JpaRepository<Parada, Long> {
 
     @Query("SELECT p FROM Parada p WHERE p.latitud = ?1 AND p.longitud = ?2")
-    Parada buscarPorLatYLong(float latitud, float longitud);
+    Parada buscarPorLatYLong(double latitud, double longitud);
 }
